refactor: fuzzy search 유틸리티와 HighlightedText 컴포넌트를 공통 모듈로 추출
