+++ src/components/FolderSearchInput.tsx
@@ -2,6 +2,8 @@
 
 import { useState, useRef, useEffect, useCallback } from "react";
 import { useTranslations } from "next-intl";
+import { fuzzyMatch, type FuzzyMatch } from "@/utils/fuzzySearch";
+import HighlightedText from "@/components/HighlightedText";
 
 interface FolderSearchInputProps {
   onSelect: (path: string) => void;
@@ -10,80 +12,6 @@ interface FolderSearchInputProps {
   placeholder?: string;
 }
 
-interface FuzzyMatch {
-  path: string;
-  score: number;
-  /** 매칭된 문자의 인덱스 배열 */
-  matchedIndices: number[];
-}
-
-/** fzf 스타일 subsequence fuzzy matching을 수행한다 */
-function fuzzyMatch(query: string, target: string): FuzzyMatch | null {
-  const lowerQuery = query.toLowerCase();
-  const lowerTarget = target.toLowerCase();
-  const matchedIndices: number[] = [];
-
-  let queryIdx = 0;
-  for (let i = 0; i < lowerTarget.length && queryIdx < lowerQuery.length; i++) {
-    if (lowerTarget[i] === lowerQuery[queryIdx]) {
-      matchedIndices.push(i);
-      queryIdx++;
-    }
-  }
-
-  if (queryIdx !== lowerQuery.length) return null;
-
-  /** 점수 계산: 연속 매칭 보너스 + 경로 끝부분(폴더명) 매칭 가중치 */
-  let score = 0;
-  const lastSlash = target.lastIndexOf("/");
-
-  for (let i = 0; i < matchedIndices.length; i++) {
-    const idx = matchedIndices[i];
-    if (idx > lastSlash) score += 2;
-    else score += 1;
-
-    if (i > 0 && matchedIndices[i] === matchedIndices[i - 1] + 1) {
-      score += 3;
-    }
-  }
-
-  return { path: target, score, matchedIndices };
-}
-
-/** 매칭된 문자를 하이라이트하여 렌더링한다 */
-function HighlightedPath({ path, matchedIndices }: { path: string; matchedIndices: number[] }) {
-  const matchSet = new Set(matchedIndices);
-  const segments: { text: string; highlighted: boolean }[] = [];
-  let current = "";
-  let currentHighlighted = false;
-
-  for (let i = 0; i < path.length; i++) {
-    const isMatch = matchSet.has(i);
-    if (i === 0) {
-      currentHighlighted = isMatch;
-      current = path[i];
-    } else if (isMatch === currentHighlighted) {
-      current += path[i];
-    } else {
-      segments.push({ text: current, highlighted: currentHighlighted });
-      current = path[i];
-      currentHighlighted = isMatch;
-    }
-  }
-  if (current) segments.push({ text: current, highlighted: currentHighlighted });
-
-  return (
-    <span className="font-mono text-xs">
-      {segments.map((seg, i) =>
-        seg.highlighted ? (
-          <span key={i} className="text-brand-primary font-bold">{seg.text}</span>
-        ) : (
-          <span key={i}>{seg.text}</span>
-        )
-      )}
-    </span>
-  );
-}
 
 export default function FolderSearchInput({
   onSelect,
@@ -266,7 +194,7 @@ export default function FolderSearchInput({
                     onClick={() => handleSelect(match.path)}
                     onMouseEnter={() => setSelectedIndex(index)}
                   >
-                    <HighlightedPath path={match.path} matchedIndices={match.matchedIndices} />
+                    <HighlightedText text={match.path} matchedIndices={match.matchedIndices} />
                   </button>
                 ))}
               </div>
